Add intercepts and mock data for offences, prosecutors, and results
- Created `OffenceIntercepts.ts` to handle interception of offences data.
- Added mock data for offences in `OffenceIntercept.mocks.ts`.
- Implemented `ProsecutorsIntercepts.ts` for intercepting prosecutors data.
- Introduced mock data for prosecutors in `ProsecutorsIntercept.mocks.ts`.
- Developed `ResultsIntercepts.ts` to manage results interception and added relevant mock data.
- Created mock data for next permitted enforcement actions in `NextPermittedEnforcementActionsIntercept.mocks.ts`.
- Updated existing tests to utilize new mock data and intercepts for better isolation and reliability.
