+++ cypress/component/CommonIntercepts/CommonIntercept.mocks.ts
@@ -1,1713 +1,9 @@
-import { IOpalFinesEnforcersRefData } from '@app/flows/fines/services/opal-fines-service/interfaces/opal-fines-enforcers-ref-data.interface';
-import { IOpalFinesCourtRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-court-ref-data.interface';
-import { IOpalFinesLocalJusticeAreaRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-local-justice-area-ref-data.interface';
-import { IOpalFinesMajorCreditorRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-major-creditor-ref-data.interface';
-import { IOpalFinesOffencesRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-offences-ref-data.interface';
-import { IOpalFinesProsecutorRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-prosecutor-ref-data.interface';
-import { IOpalFinesResultRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-result-ref-data.interface';
-import { IOpalFinesResultsRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-results-ref-data.interface';
-
-export const BUSINESS_UNIT_77_MOCK = {
-  businessUnitId: 77,
-  businessUnitName: 'Camberwell Green',
-  businessUnitCode: '0046',
-  businessUnitType: 'Accounting Division',
-  accountNumberPrefix: null,
-  opalDomain: 'Fines',
-  welshLanguage: false,
-  parentBusinessUnit: {
-    businessUnitId: 1001,
-    businessUnitName: 'Greater London',
-    businessUnitCode: '01',
-    businessUnitType: 'Area',
-    accountNumberPrefix: null,
-    opalDomain: null,
-    welshLanguage: false,
-    parentBusinessUnit: null,
-    configurationItems: [],
-  },
-  configurationItems: [],
-};
-
-export const OPAL_OFFENCE_BY_ID_MOCK = [
-  {
-    offenceId: 47883,
-    businessUnitId: null,
-    cjsCode: 'CD09019',
-    offenceTitle:
-      'Carrier - vehicle carrying dangerous / pressurised goods with an inadequacy of fire extinguishers on the vehicle',
-    offenceTitleCy: null,
-    dateUsedFrom: '2009-07-01T00:00:00',
-    dateUsedTo: null,
-    offenceOas:
-      'Contrary to regulation 5 of the Carriage of Dangerous Goods and Use of Transportable Pressure Equipment Regulations 2009 and section 33(1)(c) and (3) of the Health and Safety at Work etc Act 1974.',
-    offenceOasCy: null,
-    businessUnit: null,
-  },
-  {
-    offenceId: 33369,
-    businessUnitId: null,
-    cjsCode: 'FP12345',
-    offenceTitle: 'Fixed penalty offence - Vehicle related',
-    offenceTitleCy: null,
-    dateUsedFrom: '2020-01-01T00:00:00',
-    dateUsedTo: null,
-    offenceOas: 'Fixed penalty notice offence under relevant regulations.',
-    offenceOasCy: null,
-    businessUnit: null,
-  },
-];
-
-export const OPAL_FINES_OFFENCES_REF_DATA_MOCK: IOpalFinesOffencesRefData = {
-  count: 18,
-  refData: [
-    {
-      offence_id: 314441,
-      get_cjs_code: 'AK123456',
-      business_unit_id: 52,
-      offence_title: 'ak test',
-      offence_title_cy: null,
-      date_used_from: '2022-11-12T00:00:00',
-      date_used_to: null,
-      offence_oas: 'act 1 section 12',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 47883,
-      get_cjs_code: 'CD09019',
-      offence_title:
-        'Carrier - vehicle carrying dangerous / pressurised goods with an inadequacy of fire extinguishers on the vehicle',
-      date_used_from: '2009-07-01T00:00:00Z',
-      offence_oas:
-        'Contrary to regulation 5 of the Carriage of Dangerous Goods and Use of Transportable Pressure Equipment Regulations 2009 and section 33(1)(c) and (3) of the Health and Safety at Work etc Act 1974.',
-      business_unit_id: 77,
-      offence_title_cy: null,
-      date_used_to: null,
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 33369,
-      get_cjs_code: 'FP12345',
-      business_unit_id: 77,
-      offence_title: 'Fixed penalty offence - Vehicle related',
-      offence_title_cy: null,
-      date_used_from: '2020-01-01T00:00:00Z',
-      date_used_to: null,
-      offence_oas: 'Fixed penalty notice offence under relevant regulations.',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314683,
-      get_cjs_code: 'CA03010D',
-      business_unit_id: 22,
-      offence_title: 'No Television Licence',
-      offence_title_cy: 'WWWWWWWWWWWWWWWWWWWWWW',
-      date_used_from: '2001-01-01T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Contrary to section 363(2) and (4) of the Communications Act 2003.',
-      offence_oas_cy: 'AAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
-    },
-    {
-      offence_id: 314201,
-      get_cjs_code: 'GMMET001',
-      business_unit_id: 22,
-      offence_title: 'No Ticket on Metrolink',
-      offence_title_cy: null,
-      date_used_from: '2008-01-01T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Contrary to Regulation 4(1)(a) of the Greater Manchester Metrolink System Byelaws',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314203,
-      get_cjs_code: 'GMMET001',
-      business_unit_id: 1020,
-      offence_title: 'No Ticket on Metrolink',
-      offence_title_cy: null,
-      date_used_from: '2008-01-01T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Contrary to Regulation 4(1)(a) of the Greater Manchester Metrolink System Byelaws',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314205,
-      get_cjs_code: 'GMMET001',
-      business_unit_id: 1037,
-      offence_title: 'No Ticket on Metrolink',
-      offence_title_cy: null,
-      date_used_from: '2008-01-01T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Contrary to Regulation 4(1)(a) of the Greater Manchester Metrolink System Byelaws',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314321,
-      get_cjs_code: 'GST001',
-      business_unit_id: 52,
-      offence_title: 'GST001',
-      offence_title_cy: null,
-      date_used_from: '2010-10-10T00:00:00',
-      date_used_to: null,
-      offence_oas: 'GST001',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314481,
-      get_cjs_code: 'HU11001',
-      business_unit_id: 1042,
-      offence_title: 'Humberside By Law 001',
-      offence_title_cy: null,
-      date_used_from: '2011-01-01T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Act 1 Section 1',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314603,
-      get_cjs_code: 'JF000001',
-      business_unit_id: 52,
-      offence_title: 'jf test',
-      offence_title_cy: null,
-      date_used_from: '2012-06-19T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Act 1, Scene 1. A riverside.',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314281,
-      get_cjs_code: 'LOC001',
-      business_unit_id: 1042,
-      offence_title: 'GST Local Offence Test',
-      offence_title_cy: null,
-      date_used_from: '2010-01-01T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Act & Section',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314361,
-      get_cjs_code: 'NEWLOCOF',
-      business_unit_id: 1042,
-      offence_title: 'New Local Offence',
-      offence_title_cy: null,
-      date_used_from: '2010-03-18T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Act & Section',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314723,
-      get_cjs_code: 'NW01001',
-      business_unit_id: 45,
-      offence_title: "Fail to be long enough to fit as many words as necessary so that the title isn't truncated ",
-      offence_title_cy: null,
-      date_used_from: '2014-01-01T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Contrary to an Act',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314521,
-      get_cjs_code: 'TEST12',
-      business_unit_id: 52,
-      offence_title: 'Test',
-      offence_title_cy: null,
-      date_used_from: '2011-10-11T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Act 1 Section 12',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314561,
-      get_cjs_code: 'TEST13',
-      business_unit_id: 52,
-      offence_title: 'test 13',
-      offence_title_cy: null,
-      date_used_from: '2011-11-07T00:00:00',
-      date_used_to: null,
-      offence_oas: 'act 1 s 13',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314643,
-      get_cjs_code: 'TH68001',
-      business_unit_id: 52,
-      offence_title: 'fdfafsdsafsdf',
-      offence_title_cy: null,
-      date_used_from: '2014-10-11T00:00:00',
-      date_used_to: null,
-      offence_oas: 'fdsfasdf',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314601,
-      get_cjs_code: 'TST50001',
-      business_unit_id: 52,
-      offence_title: 'Local Offence Blocked for Use',
-      offence_title_cy: null,
-      date_used_from: '2012-07-20T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Act 5 Section 1',
-      offence_oas_cy: null,
-    },
-    {
-      offence_id: 314241,
-      get_cjs_code: 'WE99999',
-      business_unit_id: 22,
-      offence_title: 'Overtaking on a Motorway without indicating',
-      offence_title_cy: null,
-      date_used_from: '2009-06-29T00:00:00',
-      date_used_to: null,
-      offence_oas: 'Motoring Act Contrary to the Motor Act.',
-      offence_oas_cy: null,
-    },
-  ],
-};
-
-export const OPAL_FINES_LOCAL_JUSTICE_AREA_REF_DATA_MOCK: IOpalFinesLocalJusticeAreaRefData = {
-  count: 6,
-  refData: [
-    {
-      name: 'Asylum & Immigration Tribunal',
-      postcode: 'LE1 6ZX',
-      address_line_1: 'Arnhem Support Centre',
-      local_justice_area_id: 9985,
-      lja_code: '9985',
-    },
-    {
-      name: "Avon & Somerset Magistrates' Court",
-      postcode: 'BS1 3NU',
-      address_line_1: "The Magistrates' Court",
-      local_justice_area_id: 5735,
-      lja_code: '1450',
-    },
-    {
-      name: 'Aberdeen Sheriff Court District',
-      postcode: 'AB10 1WP',
-      address_line_1: 'Castle Street',
-      local_justice_area_id: 9701,
-      lja_code: '9701',
-    },
-    {
-      name: "Bedfordshire Magistrates' Court",
-      postcode: 'LU1 5BL',
-      address_line_1: 'Stuart Street',
-      local_justice_area_id: 4165,
-      lja_code: '1080',
-    },
-    {
-      name: "Berkshire Magistrates' Court",
-      postcode: 'RG1 7TQ',
-      address_line_1: 'Civic Centre',
-      local_justice_area_id: 4125,
-      lja_code: '1920',
-    },
-    {
-      name: "Birmingham and Solihull Magistrates' Court",
-      postcode: 'B4 6QA',
-      address_line_1: 'Victoria Law Courts',
-      local_justice_area_id: 5004,
-      lja_code: '2922',
-    },
-  ],
-};
-
-export const OPAL_FINES_RESULTS_REF_DATA_MOCK: IOpalFinesResultsRefData = {
-  count: 8,
-  refData: [
-    {
-      result_id: 'FCC',
-      result_title: 'Criminal Courts Charge',
-      result_title_cy: "Ffi'r Llysoedd Troseddol",
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 6,
-    },
-    {
-      result_id: 'FCOMP',
-      result_title: 'Compensation',
-      result_title_cy: 'Iawndal',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'Any',
-      imposition_allocation_order: 1,
-    },
-    {
-      result_id: 'FCOST',
-      result_title: 'Costs',
-      result_title_cy: 'Costau',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '!CPS',
-      imposition_allocation_order: 3,
-    },
-    {
-      result_id: 'FCPC',
-      result_title: 'Costs to Crown Prosecution Service',
-      result_title_cy: 'Costau i Wasanaeth Erlyn y Goron',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CPS',
-      imposition_allocation_order: 3,
-    },
-    {
-      result_id: 'FFR',
-      result_title: 'FORFEITED RECOGNISANCE',
-      result_title_cy: null,
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 10,
-    },
-    {
-      result_id: 'FO',
-      result_title: 'Fine',
-      result_title_cy: 'Dirwy',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 5,
-    },
-    {
-      result_id: 'FVEBD',
-      result_title: 'Vehicle Excise Back Duty',
-      result_title_cy: 'Ôl-dreth ar y Dreth Cerbyd',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 8,
-    },
-    {
-      result_id: 'FVS',
-      result_title: 'Victim Surcharge',
-      result_title_cy: 'Gordal Dioddefwr',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: 'CF',
-      imposition_allocation_order: 2,
-    },
-  ],
-};
-
-export const OPAL_FINES_RESULT_REF_DATA_MOCK: IOpalFinesResultRefData[] = [
-  {
-    result_id: 'REM',
-    result_title: 'Reminder of Unpaid Fine',
-    result_title_cy: 'Nodyn atgoffa terfynol am ddirwy heb ei thalu',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: false,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" }]',
-  },
-  {
-    result_id: 'FO',
-    result_title: 'Fine',
-    result_title_cy: 'Dirwy',
-    result_type: 'Result',
-    active: true,
-    imposition_allocation_priority: 6,
-    imposition_creditor: 'CF',
-    imposition: true,
-    imposition_category: 'Fines',
-    imposition_accruing: false,
-    enforcement: false,
-    enforcement_override: false,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: false,
-    generates_warrant: false,
-    result_parameters: undefined,
-    allow_payment_terms: true,
-    requires_employment_data: false,
-    allow_additional_action: false,
-    enf_next_permitted_actions: undefined,
-    requires_lja: false,
-    manual_enforcement: false,
-  },
-  {
-    result_id: 'ABDC',
-    result_title: 'Application made for Benefit Deductions',
-    result_title_cy: 'Cais am dynnu arian o fudd-daliadau',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: true,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false", "hint":"For example, LIVE UC or AMEND UC +400.00" },{ "name":"enforcer", "prompt":"Job Center Plus code (Enforcer code)", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'AEOC',
-    result_title: 'Attachment of Earnings Order - with Collection Order',
-    result_title_cy: 'Gorchymyn atafaelu enillion - cyfrif dirwy',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: true,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false", "hint":"For example, include employer name" }]',
-  },
-  {
-    result_id: 'BWTD',
-    result_title: 'Bail Warrant - dated',
-    result_title_cy: 'Gwarant Mechniaeth - Gyda dyddiad',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: true,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" },{ "name":"hearingdate", "prompt":"Adjourned to", "type":"date", "min":1, "language_dependent":false, "hint":"For example, 31/01/2023" },{ "name":"courtcode", "prompt":"Court code", "type":"integer", "min":1, "max":999, "language_dependent":false },{ "name":"enforcer", "prompt":"Warrant enforcer (Enforcer code)", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'BWTU',
-    result_title: 'Bail Warrant - undated',
-    result_title_cy: 'Gwarant Mechniaeth - Heb Ddyddiad',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" },{ "name":"baildirection", "prompt":"Date and time (Bail direction)", "type":"text", "min":0, "max":500, "language_dependent":false, "hint":"As ordered in court. For example, 31/01/2023 and 14:30" },{ "name":"courtdetails", "prompt":"Court details", "type":"text", "min":1, "max":100, "language_dependent":false },{ "name":"enforcer", "prompt":"Enforcer code", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'CLAMPO',
-    result_title: 'Clamping Order',
-    result_title_cy: '',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false", "hint":"For example, vehicle details held on account" },{ "name":"effectivedate", "prompt":"Effective from date", "type":"date", "language_dependent":true, "hint":"For example, 31/01/2023" },{ "name":"enforcer", "prompt":"Clamping contractor code (Enforcer code)", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'CWN',
-    result_title: 'Committal Warrant Notice',
-    result_title_cy: 'Rhybudd o Warant Traddodi',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: true,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false", "hint":"For example, failed to comply with suspended committal" },{ "name":"hearingdate", "prompt":"Adjourned to", "type":"date", "min":1, "language_dependent":false, "hint":"For example, 31/01/2023" },{ "name":"courtcode", "prompt":"Court code", "type":"integer", "min":1, "max":999, "language_dependent":false }]',
-  },
-  {
-    result_id: 'DW',
-    result_title: 'Warrant of Control',
-    result_title_cy: 'Gwarant Atafaelu',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    generates_warrant: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" },{ "name":"enforcer", "prompt":"Enforcer code", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'MAN',
-    result_title: 'Manual Enforcement',
-    result_title_cy: '',
-    result_type: 'Result',
-    active: false,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters: undefined,
-  },
-  {
-    result_id: 'NBWT',
-    result_title: 'No bail warrant',
-    result_title_cy: 'Gwarant dim mechniaeth',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: true,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" },{ "name":"courtdetails", "prompt":"Court details", "type":"text", "min":0, "max":100, "language_dependent":false },{ "name":"enforcer", "prompt":"Enforcer code", "type":"enforcers", "min":1, "max":1, "language_dependent":"true" }]',
-  },
-  {
-    result_id: 'REGF',
-    result_title: 'Registration of Fine',
-    result_title_cy: 'Rhybudd Cofrestru',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: true,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: true,
-    prevent_payment_card: false,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" }]',
-  },
-  {
-    result_id: 'SUMA',
-    result_title: 'Enforcement Summons (Automatic enforcement only)',
-    result_title_cy: 'Gwŷs Gorfodi',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: true,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: true,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: true,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" }]',
-  },
-  {
-    result_id: 'TFOOUT',
-    result_title: 'Transfer of Fine Order to a Court in England or Wales',
-    result_title_cy: 'Trosglwyddo Gorchymyn Dirwy i Lys yng Nghymru neu Lloegr',
-    result_type: 'Action',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: false,
-    enforcement_override: true,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: true,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: true,
-    lists_monies: false,
-    generates_warrant: undefined,
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: true,
-    manual_enforcement: undefined,
-    result_parameters: undefined,
-  },
-  {
-    result_id: 'FSN',
-    result_title: 'Further Steps Notice',
-    result_title_cy: 'Rhybudd o gamau pellach',
-    result_type: 'Result',
-    active: true,
-    imposition: false,
-    imposition_accruing: false,
-    enforcement: true,
-    enforcement_override: false,
-    further_enforcement_warn: false,
-    further_enforcement_disallow: false,
-    enforcement_hold: false,
-    requires_enforcer: false,
-    generates_hearing: false,
-    collection_order: false,
-    extend_ttp_disallow: false,
-    extend_ttp_preserve_last_enf: false,
-    prevent_payment_card: false,
-    lists_monies: false,
-    result_parameters:
-      '[{ "name":"reason", "prompt":"Reason", "type":"text", "min":1, "max":24, "language_dependent":"false" }]',
-    imposition_category: undefined,
-    imposition_allocation_priority: undefined,
-    imposition_creditor: undefined,
-    generates_warrant: undefined,
-    allow_payment_terms: undefined,
-    requires_employment_data: undefined,
-    allow_additional_action: undefined,
-    enf_next_permitted_actions: undefined,
-    requires_lja: undefined,
-    manual_enforcement: undefined,
-  },
-];
-
-export const OPAL_FINES_ENF_OVERRIDE_RESULT_REF_DATA_MOCK: IOpalFinesResultsRefData = {
-  count: 13,
-  refData: [
-    {
-      result_id: 'ABDC',
-      result_title: 'Application made for Benefit Deductions',
-      result_title_cy: 'Cais am dynnu arian o fudd-daliadau',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'AEOC',
-      result_title: 'Attachment of Earnings Order - with Collection Order',
-      result_title_cy: 'Gorchymyn atafaelu enillion - cyfrif dirwy',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'BWTD',
-      result_title: 'Bail Warrant - dated',
-      result_title_cy: 'Gwarant Mechniaeth - Gyda dyddiad',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'BWTU',
-      result_title: 'Bail Warrant - undated',
-      result_title_cy: 'Gwarant Mechniaeth - Heb Ddyddiad',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'CLAMPO',
-      result_title: 'Clamping Order',
-      result_title_cy: null,
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'CWN',
-      result_title: 'Committal Warrant Notice',
-      result_title_cy: 'Rhybudd o Warant Traddodi',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'DW',
-      result_title: 'Warrant of Control',
-      result_title_cy: 'Gwarant Atafaelu',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'MAN',
-      result_title: 'Manual Enforcement',
-      result_title_cy: null,
-      active: false,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'NBWT',
-      result_title: 'No bail warrant',
-      result_title_cy: 'Gwarant dim mechniaeth',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'REGF',
-      result_title: 'Registration of Fine',
-      result_title_cy: 'Rhybudd Cofrestru',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'SUMA',
-      result_title: 'Enforcement Summons (Automatic enforcement only)',
-      result_title_cy: 'Gwŷs Gorfodi',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'TFOOUT',
-      result_title: 'Transfer of Fine Order to a Court in England or Wales',
-      result_title_cy: 'Trosglwyddo Gorchymyn Dirwy i Lys yng Nghymru neu Lloegr',
-      active: true,
-      result_type: 'Action',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'FSN',
-      result_title: 'Further Steps Notice',
-      result_title_cy: 'Rhybudd o gamau pellach',
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-  ],
-};
-
-export const OPAL_FINES_MAJOR_CREDITOR_REF_DATA_MOCK: IOpalFinesMajorCreditorRefData = {
-  count: 4,
-  refData: [
-    {
-      account_number: null,
-      business_unit_id: 77,
-      creditor_account_id: null,
-      creditor_account_type: null,
-      from_suspense: null,
-      hold_payout: null,
-      last_changed_date: null,
-      major_creditor_code: 'AGAL',
-      major_creditor_id: 3855,
-      major_creditor_party_id: null,
-      name: 'Abellio Greater Anglia',
-      postcode: null,
-      prosecution_service: true,
-    },
-    {
-      account_number: null,
-      business_unit_id: 77,
-      creditor_account_id: null,
-      creditor_account_type: null,
-      from_suspense: null,
-      hold_payout: null,
-      last_changed_date: null,
-      major_creditor_code: 'ABJP',
-      major_creditor_id: 3854,
-      major_creditor_party_id: null,
-      name: 'Aberdeen JP Court',
-      postcode: null,
-      prosecution_service: false,
-    },
-    {
-      account_number: null,
-      business_unit_id: 77,
-      creditor_account_id: null,
-      creditor_account_type: null,
-      from_suspense: null,
-      hold_payout: null,
-      last_changed_date: null,
-      major_creditor_code: 'ALDI',
-      major_creditor_id: 3856,
-      major_creditor_party_id: null,
-      name: 'Aldi Stores Ltd',
-      postcode: null,
-      prosecution_service: false,
-    },
-    {
-      account_number: null,
-      business_unit_id: 77,
-      creditor_account_id: null,
-      creditor_account_type: null,
-      from_suspense: null,
-      hold_payout: null,
-      last_changed_date: null,
-      major_creditor_code: 'ARVA',
-      major_creditor_id: 3858,
-      major_creditor_party_id: null,
-      name: 'Arriva Rail North',
-      postcode: null,
-      prosecution_service: false,
-    },
-  ],
-};
-
-export const OPAL_FINES_COURT_REF_DATA_MOCK: IOpalFinesCourtRefData = {
-  count: 8,
-  refData: [
-    {
-      court_id: 1865,
-      business_unit_id: 61,
-      court_code: 101,
-      name: 'Historic Debt Database',
-      name_cy: null,
-      national_court_code: null,
-    },
-    {
-      court_id: 1867,
-      business_unit_id: 61,
-      court_code: 998,
-      name: 'Historic Debt Database',
-      name_cy: null,
-      national_court_code: null,
-    },
-    {
-      court_id: 1866,
-      business_unit_id: 61,
-      court_code: 102,
-      name: 'HISTORIC DEBT LODGE COURT',
-      name_cy: null,
-      national_court_code: null,
-    },
-    {
-      court_id: 1868,
-      business_unit_id: 61,
-      court_code: 999,
-      name: 'Port Talbot Justice Centre',
-      name_cy: null,
-      national_court_code: null,
-    },
-    {
-      court_id: 770000000001,
-      business_unit_id: 77,
-      court_code: 104,
-      name: 'Camberwell Green Magistrates Court',
-      name_cy: null,
-      national_court_code: null,
-    },
-    {
-      court_id: 770000000002,
-      business_unit_id: 77,
-      court_code: 105,
-      name: 'Southwark Crown Court',
-      name_cy: null,
-      national_court_code: null,
-    },
-    {
-      court_id: 770000000003,
-      business_unit_id: 77,
-      court_code: 106,
-      name: 'Inner London Crown Court',
-      name_cy: null,
-      national_court_code: null,
-    },
-    {
-      court_id: 770000000004,
-      business_unit_id: 77,
-      court_code: 107,
-      name: 'Woolwich Crown Court',
-      name_cy: null,
-      national_court_code: null,
-    },
-  ],
-};
-export const OPAL_FINES_PROSECUTOR_REF_DATA_MOCK: IOpalFinesProsecutorRefData = {
-  count: 4,
-  ref_data: [
-    {
-      prosecutor_id: 1865,
-      prosecutor_name: 'Central ticket office',
-      prosecutor_code: '998',
-      address_line_1: null,
-      address_line_2: null,
-      address_line_3: null,
-      address_line_4: null,
-      address_line_5: null,
-      postcode: null,
-      end_date: null,
-    },
-    {
-      prosecutor_id: 1223,
-      prosecutor_name: 'Police force',
-      prosecutor_code: '123',
-      address_line_1: null,
-      address_line_2: null,
-      address_line_3: null,
-      address_line_4: null,
-      address_line_5: null,
-      postcode: null,
-      end_date: null,
-    },
-    {
-      prosecutor_id: 3523,
-      prosecutor_name: 'Other',
-      prosecutor_code: '433',
-      address_line_1: null,
-      address_line_2: null,
-      address_line_3: null,
-      address_line_4: null,
-      address_line_5: null,
-      postcode: null,
-      end_date: null,
-    },
-  ],
-};
-
-export const OPAL_FINES_ENFORCER_REF_DATA_MOCK: IOpalFinesEnforcersRefData = {
-  count: 83,
-  refData: [
-    {
-      enforcer_id: 780000000002,
-      enforcer_code: 3,
-      name: '00000',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 650000000001,
-      enforcer_code: 840,
-      name: 'ATCMTest',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 470000000001,
-      enforcer_code: 101,
-      name: 'BR.Enfone',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 470000000003,
-      enforcer_code: 103,
-      name: 'BR.Enfthree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 470000000002,
-      enforcer_code: 102,
-      name: 'BR.Enftwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 50000000001,
-      enforcer_code: 101,
-      name: 'Cam.Enfone',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 50000000003,
-      enforcer_code: 103,
-      name: 'Cam.Enfthree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 50000000002,
-      enforcer_code: 102,
-      name: 'Cam.Enftwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1300000000002,
-      enforcer_code: 1,
-      name: 'CIVILIAN ENFORCEMENT OFFICERS',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 600000000001,
-      enforcer_code: 101,
-      name: 'Dyfed Enforcer 1',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1060000000041,
-      enforcer_code: 801,
-      name: 'EnfAE01',
-      name_cy: 'EnfAE01',
-    },
-    {
-      enforcer_id: 1060000000043,
-      enforcer_code: 803,
-      name: 'EnfAE02',
-      name_cy: 'EnfAE02',
-    },
-    {
-      enforcer_id: 1060000000045,
-      enforcer_code: 802,
-      name: 'EnfAE03',
-      name_cy: 'EnfAE03',
-    },
-    {
-      enforcer_id: 1060000000047,
-      enforcer_code: 804,
-      name: 'EnfAE04',
-      name_cy: 'EnfAE04',
-    },
-    {
-      enforcer_id: 1060000000049,
-      enforcer_code: 805,
-      name: 'EnfAE05',
-      name_cy: 'EnfAE05',
-    },
-    {
-      enforcer_id: 780000000001,
-      enforcer_code: 2,
-      name: 'Enfone Add3',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 600000000021,
-      enforcer_code: 901,
-      name: 'Enforcement(Dy) - ABDC',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 600000000022,
-      enforcer_code: 902,
-      name: 'Enforcement(Dy) - BWTD',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 600000000023,
-      enforcer_code: 903,
-      name: 'Enforcement(Dy) - BWTU',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 600000000024,
-      enforcer_code: 904,
-      name: 'Enforcement(Dy) - CLAMPO',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 600000000025,
-      enforcer_code: 905,
-      name: 'Enforcement(Dy) - DW',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 600000000026,
-      enforcer_code: 906,
-      name: 'Enforcement(Dy) - NBWT',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000041,
-      enforcer_code: 901,
-      name: 'Enforcement(Gw) - ABDC',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000042,
-      enforcer_code: 902,
-      name: 'Enforcement(Gw) - BWTD',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000043,
-      enforcer_code: 903,
-      name: 'Enforcement(Gw) - BWTU',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000044,
-      enforcer_code: 904,
-      name: 'Enforcement(Gw) - CLAMPO',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000045,
-      enforcer_code: 905,
-      name: 'Enforcement(Gw) - DW',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000046,
-      enforcer_code: 906,
-      name: 'Enforcement(Gw) - NBWT',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1060000000021,
-      enforcer_code: 901,
-      name: 'Enforcement(Nw) - ABDC',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1060000000022,
-      enforcer_code: 902,
-      name: 'Enforcement(Nw) - BWTD',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1060000000023,
-      enforcer_code: 903,
-      name: 'Enforcement(Nw) - BWTU',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1060000000024,
-      enforcer_code: 904,
-      name: 'Enforcement(Nw) - CLAMPO',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1060000000025,
-      enforcer_code: 905,
-      name: 'Enforcement(Nw) - DW',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1060000000026,
-      enforcer_code: 906,
-      name: 'Enforcement(Nw) - NBWT',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000041,
-      enforcer_code: 901,
-      name: 'Enforcement(SW-SMG) - ABDC',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000042,
-      enforcer_code: 902,
-      name: 'Enforcement(SW-SMG) - BWTD',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000043,
-      enforcer_code: 903,
-      name: 'Enforcement(SW-SMG) - BWTU',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000044,
-      enforcer_code: 904,
-      name: 'Enforcement(SW-SMG) - CLAMPO',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000045,
-      enforcer_code: 905,
-      name: 'Enforcement(SW-SMG) - DW',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000046,
-      enforcer_code: 906,
-      name: 'Enforcement(SW-SMG) - NBWT',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 730000000021,
-      enforcer_code: 104,
-      name: 'EnforcerFour',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 800000000001,
-      enforcer_code: 101,
-      name: 'EnforcerOne',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 730000000001,
-      enforcer_code: 101,
-      name: 'EnforcerOne',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 800000000003,
-      enforcer_code: 103,
-      name: 'EnforcerThree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 730000000003,
-      enforcer_code: 103,
-      name: 'EnforcerThree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 800000000002,
-      enforcer_code: 102,
-      name: 'EnforcerTwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 730000000002,
-      enforcer_code: 102,
-      name: 'EnforcerTwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 820000000001,
-      enforcer_code: 101,
-      name: 'Gm.Enfone',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 820000000004,
-      enforcer_code: 1,
-      name: 'Gm.Enfone Add2',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 820000000003,
-      enforcer_code: 103,
-      name: 'Gm.Enfthree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 820000000002,
-      enforcer_code: 102,
-      name: 'Gm.Enftwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 820000000005,
-      enforcer_code: 2,
-      name: 'Gm.Enftwo Add3',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000001,
-      enforcer_code: 101,
-      name: 'Gwent.Enfone',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000003,
-      enforcer_code: 103,
-      name: 'Gwent.Enfthree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000002,
-      enforcer_code: 102,
-      name: 'Gwent.Enftwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 300000000001,
-      enforcer_code: 101,
-      name: 'GY.Enfone',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 300000000003,
-      enforcer_code: 103,
-      name: 'GY.Enfthree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 300000000002,
-      enforcer_code: 102,
-      name: 'GY.Enftwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 260000000042,
-      enforcer_code: 105,
-      name: 'Job Centre Hemel Hempstead',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 260000000041,
-      enforcer_code: 104,
-      name: 'Job Centre Watford',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 260000000021,
-      enforcer_code: 101,
-      name: 'NE.Enfone',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 260000000023,
-      enforcer_code: 103,
-      name: 'NE.Enfthree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 260000000022,
-      enforcer_code: 102,
-      name: 'NE.Enftwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1060000000001,
-      enforcer_code: 101,
-      name: 'North Wales Enforcer 1',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1060000000002,
-      enforcer_code: 102,
-      name: 'North Wales Job Centre 1',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000021,
-      enforcer_code: 1,
-      name: 'NW.Enfone Add2',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 890000000022,
-      enforcer_code: 2,
-      name: 'NW.Enftwo Add3',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 570000000001,
-      enforcer_code: 101,
-      name: 'PD.EnfOne',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 570000000003,
-      enforcer_code: 103,
-      name: 'PD.EnfThree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 570000000002,
-      enforcer_code: 102,
-      name: 'PD.EnfTwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1030000000001,
-      enforcer_code: 101,
-      name: 'PE Enfone',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1030000000003,
-      enforcer_code: 103,
-      name: 'PE.Enfthree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 1030000000002,
-      enforcer_code: 102,
-      name: 'PE.Enftwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 260000000043,
-      enforcer_code: 106,
-      name: 'Police Hertfordshire',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000001,
-      enforcer_code: 101,
-      name: 'SouthWales.Enfone',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000003,
-      enforcer_code: 103,
-      name: 'SouthWales.Enfthree',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000002,
-      enforcer_code: 102,
-      name: 'SouthWales.Enftwo',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 360000000021,
-      enforcer_code: 104,
-      name: 'South Wales Job Centre',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 770000000001,
-      enforcer_code: 1,
-      name: 'The Bailiffs',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 770000000003,
-      enforcer_code: 3,
-      name: 'The DWP',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 260000000001,
-      enforcer_code: 123,
-      name: 'The enforcers name',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 770000000002,
-      enforcer_code: 2,
-      name: 'The police',
-      name_cy: null,
-    },
-    {
-      enforcer_id: 780000000021,
-      enforcer_code: 101,
-      name: 'Warrant enforcer',
-      name_cy: null,
-    },
-  ],
-};
-export const OPAL_FINES_NEXT_PERMITTED_ENFORCEMENT_ACTIONS_MOCK: IOpalFinesResultsRefData = {
-  count: 2,
-  refData: [
-    {
-      result_id: 'WOC',
-      result_title: 'Warrant of Control',
-      result_title_cy: null,
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-    {
-      result_id: 'WOA',
-      result_title: 'Warrant of Arrest',
-      result_title_cy: null,
-      active: true,
-      result_type: 'Result',
-      imposition_creditor: '',
-      imposition_allocation_order: null,
-    },
-  ],
-};
+export * from './referenceData/businessUnits/BusinessUnit.mocks';
+export * from './referenceData/courts/CourtsIntercept.mocks';
+export * from './referenceData/enforcers/EnforcersIntercept.mocks';
+export * from './referenceData/localJusticeAreas/LocalJusticeAreasIntercept.mocks';
+export * from './referenceData/majorCreditors/MajorCreditorsIntercept.mocks';
+export * from './referenceData/offences/OffenceIntercept.mocks';
+export * from './referenceData/prosecutors/ProsecutorsIntercept.mocks';
+export * from './referenceData/results/NextPermittedEnforcementActionsIntercept.mocks';
+export * from './referenceData/results/ResultsIntercept.mocks';
