+++ cypress/component/CommonIntercepts/compositions/ReviewAccountIntercepts.ts
@@ -1,0 +1,15 @@
+import { interceptCourtsByBU } from '../referenceData/courts/CourtsIntercepts';
+import { interceptLocalJusticeAreas } from '../referenceData/localJusticeAreas/LocalJusticeAreasIntercepts';
+import { interceptMajorCreditorsByBU } from '../referenceData/majorCreditors/MajorCreditorsIntercepts';
+import { interceptProsecutors } from '../referenceData/prosecutors/ProsecutorsIntercepts';
+import { interceptResultsByIds } from '../referenceData/results/ResultsIntercepts';
+
+const REVIEW_ACCOUNT_RESULT_IDS = ['FCOMP', 'FVS', 'FCOST', 'FCPC', 'FO', 'FCC', 'FVEBD', 'FFR'];
+
+export function interceptRefDataForReviewAccount(businessUnitId: number) {
+  interceptCourtsByBU(businessUnitId);
+  interceptMajorCreditorsByBU(businessUnitId);
+  interceptLocalJusticeAreas();
+  interceptResultsByIds(REVIEW_ACCOUNT_RESULT_IDS);
+  interceptProsecutors();
+}
