+++ cypress/component/CommonIntercepts/referenceData/enforcers/EnforcersIntercept.mocks.ts
@@ -1,0 +1,505 @@
+import { IOpalFinesEnforcersRefData } from '@app/flows/fines/services/opal-fines-service/interfaces/opal-fines-enforcers-ref-data.interface';
+
+export const OPAL_FINES_ENFORCER_REF_DATA_MOCK: IOpalFinesEnforcersRefData = {
+  count: 83,
+  refData: [
+    {
+      enforcer_id: 780000000002,
+      enforcer_code: 3,
+      name: '00000',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 650000000001,
+      enforcer_code: 840,
+      name: 'ATCMTest',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 470000000001,
+      enforcer_code: 101,
+      name: 'BR.Enfone',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 470000000003,
+      enforcer_code: 103,
+      name: 'BR.Enfthree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 470000000002,
+      enforcer_code: 102,
+      name: 'BR.Enftwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 50000000001,
+      enforcer_code: 101,
+      name: 'Cam.Enfone',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 50000000003,
+      enforcer_code: 103,
+      name: 'Cam.Enfthree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 50000000002,
+      enforcer_code: 102,
+      name: 'Cam.Enftwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1300000000002,
+      enforcer_code: 1,
+      name: 'CIVILIAN ENFORCEMENT OFFICERS',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 600000000001,
+      enforcer_code: 101,
+      name: 'Dyfed Enforcer 1',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1060000000041,
+      enforcer_code: 801,
+      name: 'EnfAE01',
+      name_cy: 'EnfAE01',
+    },
+    {
+      enforcer_id: 1060000000043,
+      enforcer_code: 803,
+      name: 'EnfAE02',
+      name_cy: 'EnfAE02',
+    },
+    {
+      enforcer_id: 1060000000045,
+      enforcer_code: 802,
+      name: 'EnfAE03',
+      name_cy: 'EnfAE03',
+    },
+    {
+      enforcer_id: 1060000000047,
+      enforcer_code: 804,
+      name: 'EnfAE04',
+      name_cy: 'EnfAE04',
+    },
+    {
+      enforcer_id: 1060000000049,
+      enforcer_code: 805,
+      name: 'EnfAE05',
+      name_cy: 'EnfAE05',
+    },
+    {
+      enforcer_id: 780000000001,
+      enforcer_code: 2,
+      name: 'Enfone Add3',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 600000000021,
+      enforcer_code: 901,
+      name: 'Enforcement(Dy) - ABDC',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 600000000022,
+      enforcer_code: 902,
+      name: 'Enforcement(Dy) - BWTD',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 600000000023,
+      enforcer_code: 903,
+      name: 'Enforcement(Dy) - BWTU',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 600000000024,
+      enforcer_code: 904,
+      name: 'Enforcement(Dy) - CLAMPO',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 600000000025,
+      enforcer_code: 905,
+      name: 'Enforcement(Dy) - DW',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 600000000026,
+      enforcer_code: 906,
+      name: 'Enforcement(Dy) - NBWT',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000041,
+      enforcer_code: 901,
+      name: 'Enforcement(Gw) - ABDC',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000042,
+      enforcer_code: 902,
+      name: 'Enforcement(Gw) - BWTD',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000043,
+      enforcer_code: 903,
+      name: 'Enforcement(Gw) - BWTU',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000044,
+      enforcer_code: 904,
+      name: 'Enforcement(Gw) - CLAMPO',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000045,
+      enforcer_code: 905,
+      name: 'Enforcement(Gw) - DW',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000046,
+      enforcer_code: 906,
+      name: 'Enforcement(Gw) - NBWT',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1060000000021,
+      enforcer_code: 901,
+      name: 'Enforcement(Nw) - ABDC',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1060000000022,
+      enforcer_code: 902,
+      name: 'Enforcement(Nw) - BWTD',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1060000000023,
+      enforcer_code: 903,
+      name: 'Enforcement(Nw) - BWTU',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1060000000024,
+      enforcer_code: 904,
+      name: 'Enforcement(Nw) - CLAMPO',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1060000000025,
+      enforcer_code: 905,
+      name: 'Enforcement(Nw) - DW',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1060000000026,
+      enforcer_code: 906,
+      name: 'Enforcement(Nw) - NBWT',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000041,
+      enforcer_code: 901,
+      name: 'Enforcement(SW-SMG) - ABDC',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000042,
+      enforcer_code: 902,
+      name: 'Enforcement(SW-SMG) - BWTD',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000043,
+      enforcer_code: 903,
+      name: 'Enforcement(SW-SMG) - BWTU',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000044,
+      enforcer_code: 904,
+      name: 'Enforcement(SW-SMG) - CLAMPO',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000045,
+      enforcer_code: 905,
+      name: 'Enforcement(SW-SMG) - DW',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000046,
+      enforcer_code: 906,
+      name: 'Enforcement(SW-SMG) - NBWT',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 730000000021,
+      enforcer_code: 104,
+      name: 'EnforcerFour',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 800000000001,
+      enforcer_code: 101,
+      name: 'EnforcerOne',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 730000000001,
+      enforcer_code: 101,
+      name: 'EnforcerOne',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 800000000003,
+      enforcer_code: 103,
+      name: 'EnforcerThree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 730000000003,
+      enforcer_code: 103,
+      name: 'EnforcerThree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 800000000002,
+      enforcer_code: 102,
+      name: 'EnforcerTwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 730000000002,
+      enforcer_code: 102,
+      name: 'EnforcerTwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 820000000001,
+      enforcer_code: 101,
+      name: 'Gm.Enfone',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 820000000004,
+      enforcer_code: 1,
+      name: 'Gm.Enfone Add2',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 820000000003,
+      enforcer_code: 103,
+      name: 'Gm.Enfthree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 820000000002,
+      enforcer_code: 102,
+      name: 'Gm.Enftwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 820000000005,
+      enforcer_code: 2,
+      name: 'Gm.Enftwo Add3',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000001,
+      enforcer_code: 101,
+      name: 'Gwent.Enfone',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000003,
+      enforcer_code: 103,
+      name: 'Gwent.Enfthree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000002,
+      enforcer_code: 102,
+      name: 'Gwent.Enftwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 300000000001,
+      enforcer_code: 101,
+      name: 'GY.Enfone',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 300000000003,
+      enforcer_code: 103,
+      name: 'GY.Enfthree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 300000000002,
+      enforcer_code: 102,
+      name: 'GY.Enftwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 260000000042,
+      enforcer_code: 105,
+      name: 'Job Centre Hemel Hempstead',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 260000000041,
+      enforcer_code: 104,
+      name: 'Job Centre Watford',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 260000000021,
+      enforcer_code: 101,
+      name: 'NE.Enfone',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 260000000023,
+      enforcer_code: 103,
+      name: 'NE.Enfthree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 260000000022,
+      enforcer_code: 102,
+      name: 'NE.Enftwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1060000000001,
+      enforcer_code: 101,
+      name: 'North Wales Enforcer 1',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1060000000002,
+      enforcer_code: 102,
+      name: 'North Wales Job Centre 1',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000021,
+      enforcer_code: 1,
+      name: 'NW.Enfone Add2',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 890000000022,
+      enforcer_code: 2,
+      name: 'NW.Enftwo Add3',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 570000000001,
+      enforcer_code: 101,
+      name: 'PD.EnfOne',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 570000000003,
+      enforcer_code: 103,
+      name: 'PD.EnfThree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 570000000002,
+      enforcer_code: 102,
+      name: 'PD.EnfTwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1030000000001,
+      enforcer_code: 101,
+      name: 'PE Enfone',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1030000000003,
+      enforcer_code: 103,
+      name: 'PE.Enfthree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 1030000000002,
+      enforcer_code: 102,
+      name: 'PE.Enftwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 260000000043,
+      enforcer_code: 106,
+      name: 'Police Hertfordshire',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000001,
+      enforcer_code: 101,
+      name: 'SouthWales.Enfone',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000003,
+      enforcer_code: 103,
+      name: 'SouthWales.Enfthree',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000002,
+      enforcer_code: 102,
+      name: 'SouthWales.Enftwo',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 360000000021,
+      enforcer_code: 104,
+      name: 'South Wales Job Centre',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 770000000001,
+      enforcer_code: 1,
+      name: 'The Bailiffs',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 770000000003,
+      enforcer_code: 3,
+      name: 'The DWP',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 260000000001,
+      enforcer_code: 123,
+      name: 'The enforcers name',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 770000000002,
+      enforcer_code: 2,
+      name: 'The police',
+      name_cy: null,
+    },
+    {
+      enforcer_id: 780000000021,
+      enforcer_code: 101,
+      name: 'Warrant enforcer',
+      name_cy: null,
+    },
+  ],
+};
