+++ cypress/component/CommonIntercepts/referenceData/majorCreditors/MajorCreditorsIntercept.mocks.ts
@@ -1,0 +1,67 @@
+import { IOpalFinesMajorCreditorRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-major-creditor-ref-data.interface';
+
+export const OPAL_FINES_MAJOR_CREDITOR_REF_DATA_MOCK: IOpalFinesMajorCreditorRefData = {
+  count: 4,
+  refData: [
+    {
+      account_number: null,
+      business_unit_id: 77,
+      creditor_account_id: null,
+      creditor_account_type: null,
+      from_suspense: null,
+      hold_payout: null,
+      last_changed_date: null,
+      major_creditor_code: 'AGAL',
+      major_creditor_id: 3855,
+      major_creditor_party_id: null,
+      name: 'Abellio Greater Anglia',
+      postcode: null,
+      prosecution_service: true,
+    },
+    {
+      account_number: null,
+      business_unit_id: 77,
+      creditor_account_id: null,
+      creditor_account_type: null,
+      from_suspense: null,
+      hold_payout: null,
+      last_changed_date: null,
+      major_creditor_code: 'ABJP',
+      major_creditor_id: 3854,
+      major_creditor_party_id: null,
+      name: 'Aberdeen JP Court',
+      postcode: null,
+      prosecution_service: false,
+    },
+    {
+      account_number: null,
+      business_unit_id: 77,
+      creditor_account_id: null,
+      creditor_account_type: null,
+      from_suspense: null,
+      hold_payout: null,
+      last_changed_date: null,
+      major_creditor_code: 'ALDI',
+      major_creditor_id: 3856,
+      major_creditor_party_id: null,
+      name: 'Aldi Stores Ltd',
+      postcode: null,
+      prosecution_service: false,
+    },
+    {
+      account_number: null,
+      business_unit_id: 77,
+      creditor_account_id: null,
+      creditor_account_type: null,
+      from_suspense: null,
+      hold_payout: null,
+      last_changed_date: null,
+      major_creditor_code: 'ARVA',
+      major_creditor_id: 3858,
+      major_creditor_party_id: null,
+      name: 'Arriva Rail North',
+      postcode: null,
+      prosecution_service: false,
+    },
+  ],
+};
