+++ cypress/component/CommonIntercepts/referenceData/offences/OffenceIntercept.mocks.ts
@@ -1,0 +1,236 @@
+import { IOpalFinesOffencesRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-offences-ref-data.interface';
+
+export const OPAL_OFFENCE_BY_ID_MOCK = [
+  {
+    offenceId: 47883,
+    businessUnitId: null,
+    cjsCode: 'CD09019',
+    offenceTitle:
+      'Carrier - vehicle carrying dangerous / pressurised goods with an inadequacy of fire extinguishers on the vehicle',
+    offenceTitleCy: null,
+    dateUsedFrom: '2009-07-01T00:00:00',
+    dateUsedTo: null,
+    offenceOas:
+      'Contrary to regulation 5 of the Carriage of Dangerous Goods and Use of Transportable Pressure Equipment Regulations 2009 and section 33(1)(c) and (3) of the Health and Safety at Work etc Act 1974.',
+    offenceOasCy: null,
+    businessUnit: null,
+  },
+  {
+    offenceId: 33369,
+    businessUnitId: null,
+    cjsCode: 'FP12345',
+    offenceTitle: 'Fixed penalty offence - Vehicle related',
+    offenceTitleCy: null,
+    dateUsedFrom: '2020-01-01T00:00:00',
+    dateUsedTo: null,
+    offenceOas: 'Fixed penalty notice offence under relevant regulations.',
+    offenceOasCy: null,
+    businessUnit: null,
+  },
+];
+
+export const OPAL_FINES_OFFENCES_REF_DATA_MOCK: IOpalFinesOffencesRefData = {
+  count: 18,
+  refData: [
+    {
+      offence_id: 314441,
+      get_cjs_code: 'AK123456',
+      business_unit_id: 52,
+      offence_title: 'ak test',
+      offence_title_cy: null,
+      date_used_from: '2022-11-12T00:00:00',
+      date_used_to: null,
+      offence_oas: 'act 1 section 12',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 47883,
+      get_cjs_code: 'CD09019',
+      offence_title:
+        'Carrier - vehicle carrying dangerous / pressurised goods with an inadequacy of fire extinguishers on the vehicle',
+      date_used_from: '2009-07-01T00:00:00Z',
+      offence_oas:
+        'Contrary to regulation 5 of the Carriage of Dangerous Goods and Use of Transportable Pressure Equipment Regulations 2009 and section 33(1)(c) and (3) of the Health and Safety at Work etc Act 1974.',
+      business_unit_id: 77,
+      offence_title_cy: null,
+      date_used_to: null,
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 33369,
+      get_cjs_code: 'FP12345',
+      business_unit_id: 77,
+      offence_title: 'Fixed penalty offence - Vehicle related',
+      offence_title_cy: null,
+      date_used_from: '2020-01-01T00:00:00Z',
+      date_used_to: null,
+      offence_oas: 'Fixed penalty notice offence under relevant regulations.',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314683,
+      get_cjs_code: 'CA03010D',
+      business_unit_id: 22,
+      offence_title: 'No Television Licence',
+      offence_title_cy: 'WWWWWWWWWWWWWWWWWWWWWW',
+      date_used_from: '2001-01-01T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Contrary to section 363(2) and (4) of the Communications Act 2003.',
+      offence_oas_cy: 'AAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
+    },
+    {
+      offence_id: 314201,
+      get_cjs_code: 'GMMET001',
+      business_unit_id: 22,
+      offence_title: 'No Ticket on Metrolink',
+      offence_title_cy: null,
+      date_used_from: '2008-01-01T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Contrary to Regulation 4(1)(a) of the Greater Manchester Metrolink System Byelaws',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314203,
+      get_cjs_code: 'GMMET001',
+      business_unit_id: 1020,
+      offence_title: 'No Ticket on Metrolink',
+      offence_title_cy: null,
+      date_used_from: '2008-01-01T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Contrary to Regulation 4(1)(a) of the Greater Manchester Metrolink System Byelaws',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314205,
+      get_cjs_code: 'GMMET001',
+      business_unit_id: 1037,
+      offence_title: 'No Ticket on Metrolink',
+      offence_title_cy: null,
+      date_used_from: '2008-01-01T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Contrary to Regulation 4(1)(a) of the Greater Manchester Metrolink System Byelaws',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314321,
+      get_cjs_code: 'GST001',
+      business_unit_id: 52,
+      offence_title: 'GST001',
+      offence_title_cy: null,
+      date_used_from: '2010-10-10T00:00:00',
+      date_used_to: null,
+      offence_oas: 'GST001',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314481,
+      get_cjs_code: 'HU11001',
+      business_unit_id: 1042,
+      offence_title: 'Humberside By Law 001',
+      offence_title_cy: null,
+      date_used_from: '2011-01-01T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Act 1 Section 1',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314603,
+      get_cjs_code: 'JF000001',
+      business_unit_id: 52,
+      offence_title: 'jf test',
+      offence_title_cy: null,
+      date_used_from: '2012-06-19T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Act 1, Scene 1. A riverside.',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314281,
+      get_cjs_code: 'LOC001',
+      business_unit_id: 1042,
+      offence_title: 'GST Local Offence Test',
+      offence_title_cy: null,
+      date_used_from: '2010-01-01T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Act & Section',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314361,
+      get_cjs_code: 'NEWLOCOF',
+      business_unit_id: 1042,
+      offence_title: 'New Local Offence',
+      offence_title_cy: null,
+      date_used_from: '2010-03-18T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Act & Section',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314723,
+      get_cjs_code: 'NW01001',
+      business_unit_id: 45,
+      offence_title: "Fail to be long enough to fit as many words as necessary so that the title isn't truncated ",
+      offence_title_cy: null,
+      date_used_from: '2014-01-01T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Contrary to an Act',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314521,
+      get_cjs_code: 'TEST12',
+      business_unit_id: 52,
+      offence_title: 'Test',
+      offence_title_cy: null,
+      date_used_from: '2011-10-11T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Act 1 Section 12',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314561,
+      get_cjs_code: 'TEST13',
+      business_unit_id: 52,
+      offence_title: 'test 13',
+      offence_title_cy: null,
+      date_used_from: '2011-11-07T00:00:00',
+      date_used_to: null,
+      offence_oas: 'act 1 s 13',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314643,
+      get_cjs_code: 'TH68001',
+      business_unit_id: 52,
+      offence_title: 'fdfafsdsafsdf',
+      offence_title_cy: null,
+      date_used_from: '2014-10-11T00:00:00',
+      date_used_to: null,
+      offence_oas: 'fdsfasdf',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314601,
+      get_cjs_code: 'TST50001',
+      business_unit_id: 52,
+      offence_title: 'Local Offence Blocked for Use',
+      offence_title_cy: null,
+      date_used_from: '2012-07-20T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Act 5 Section 1',
+      offence_oas_cy: null,
+    },
+    {
+      offence_id: 314241,
+      get_cjs_code: 'WE99999',
+      business_unit_id: 22,
+      offence_title: 'Overtaking on a Motorway without indicating',
+      offence_title_cy: null,
+      date_used_from: '2009-06-29T00:00:00',
+      date_used_to: null,
+      offence_oas: 'Motoring Act Contrary to the Motor Act.',
+      offence_oas_cy: null,
+    },
+  ],
+};
