+++ cypress/component/CommonIntercepts/referenceData/prosecutors/ProsecutorsIntercept.mocks.ts
@@ -1,0 +1,43 @@
+import { IOpalFinesProsecutorRefData } from '@services/fines/opal-fines-service/interfaces/opal-fines-prosecutor-ref-data.interface';
+
+export const OPAL_FINES_PROSECUTOR_REF_DATA_MOCK: IOpalFinesProsecutorRefData = {
+  count: 4,
+  ref_data: [
+    {
+      prosecutor_id: 1865,
+      prosecutor_name: 'Central ticket office',
+      prosecutor_code: '998',
+      address_line_1: null,
+      address_line_2: null,
+      address_line_3: null,
+      address_line_4: null,
+      address_line_5: null,
+      postcode: null,
+      end_date: null,
+    },
+    {
+      prosecutor_id: 1223,
+      prosecutor_name: 'Police force',
+      prosecutor_code: '123',
+      address_line_1: null,
+      address_line_2: null,
+      address_line_3: null,
+      address_line_4: null,
+      address_line_5: null,
+      postcode: null,
+      end_date: null,
+    },
+    {
+      prosecutor_id: 3523,
+      prosecutor_name: 'Other',
+      prosecutor_code: '433',
+      address_line_1: null,
+      address_line_2: null,
+      address_line_3: null,
+      address_line_4: null,
+      address_line_5: null,
+      postcode: null,
+      end_date: null,
+    },
+  ],
+};
